% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Get all data regarding one step in the criminal process}
\usage{
get_all_data(dataset_name)
}
\arguments{
\item{dataset_name}{The name of the dataset to be retrieved.
One of 'intake', 'initiation', 'dispositions', 'sentencing'.}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Get all data regarding one step in the criminal process
}
\examples{
\dontrun{
sentencing_data <- get_all_data('sentencing')
}
}
\seealso{
\code{\link{intake}} \code{\link{initiation}} \code{\link{dispositions}} \code{\link{sentencing}}
}
