% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiation.R
\docType{data}
\name{initiation}
\alias{initiation}
\title{Initiation Data}
\format{The data frame \code{intake} has 73,029 rows and 27 variables:
\describe{
\item{case_id}{Internal unique identifier for each case}
\item{case_participant_id}{Internal unique identifier for each person associated with a case}
\item{offense_category}{Broad offense categories before specific charges are filed on a case}
\item{primary_charge}{A flag for the top charge, usually the way the case is referred to}
\item{charge_id}{Internal unique identifier for each charge filed}
\item{charge_version_id}{Internal unique identifier for each version of a charge associated with charges filed}
\item{charge_offense_title}{The offense title of the charged offense}
\item{chapter}{The legal chapter for the charge}
\item{act}{The legal act for the charge}
\item{section}{The legal section for the charge}
\item{class}{The legal class for the charge}
\item{aoic}{Administrative Office of the Illinois Courts ID for law of the charge}
\item{event}{The way the charge was brought about}
\item{event_date}{The date the charges were brought about}
\item{age_at_incident}{Recorded age at the time of the incident}
\item{gender}{Recorded gender of the defendant }
\item{race}{Recorded race of the defendant}
\item{incident_begin_date}{Date of when the incident began}
\item{incident_end_date}{Date of when the incident ended (this will be blank for incidents that did not go more than one day)}
\item{arrest_date}{Date and time of arrest}
\item{law_enforcement_agency}{Law enforcement agency associated with the arrest}
\item{unit}{The law enforcement unit associated with the arrest}
\item{incident_city}{The city where the incident took place}
\item{received_date}{Date when felony review received the case}
\item{arraignment_date}{Date of the arraignment}
\item{update_offense_category}{This field is the offense category for the case updated based upon the top charge for the primary offender. It can differ from the first offense category assigned to the case in part because cases evolve.}
\item{charge_count}{The charge count of the charged offense}
}}
\source{
\url{https://datacatalog.cookcountyil.gov/Courts/Initiation/7mck-ehwz}
}
\usage{
initiation
}
\description{
The Initiation results data presented here reflects all of the arrests that came through
the door of the State's Attorneys Office (SAO). An initiation is how an arrest turns into
a “case” in the courts. Most cases are initiated through a process known as felony review,
in which SAO attorneys make a decision whether or not to prosecute. Cases may also be
indicted by a grand jury or, in narcotics cases, filed directly by law enforcement
(labeled "BOND SET (Narcotics)" in this data). Included in this data set are the defendant
counts by initiation and year. This data includes felony cases handled by the Criminal, Narcotics,
and Special Prosecution Bureaus. It does not include information about cases processed through the
Juvenile Justice and Civil Actions Bureaus.
}
\examples{
str(initiation)
head(initiation)
summary(initiation)

}
\seealso{
\code{\link{intake}}
}
\keyword{datasets}
