% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intake.R
\docType{data}
\name{intake}
\alias{intake}
\title{Intake Data}
\format{The main data frame \code{intake} has 36,190 rows and 17 variables:
\describe{
\item{case_id}{Id}
\item{case_participant_id}{Id}
\item{received_date}{id}
}

The supplemental data frame \code{\link{initiation}} was not filtered.}
\source{
\url{https://github.com}
}
\usage{
intake
}
\description{
Intake Data
}
\examples{
str(intake)
head(intake)
summary(intake)

}
\seealso{
\code{\link{initiation}}
}
\keyword{datasets}
